+++ akshare/__init__.py
@@ -2032,9 +2032,13 @@ amac_manager_cancelled_info # 中国证券投资基金业协会-信息公示-诚
 1.6.11 add: add stock_sz_a_spot_em interface
 1.6.12 add: add stock_bj_a_spot_em interface
 1.6.13 add: add stock_new_a_spot_em interface
+1.6.14 fix: fix stock_hsgt_board_rank_em interface
+1.6.15 fix: fix bond_zh_cov_info interface
+1.6.16 fix: fix baidu_search_index interface
+1.6.17 fix: fix stock_new_a_spot_em interface
 """
 
-__version__ = "1.6.13"
+__version__ = "1.6.17"
 __author__ = "AKFamily"
 
 import sys