+++ akshare/index/index_baidu.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 # -*- coding:utf-8 -*-
 """
-Date: 2020/9/28 13:53
+Date: 2022/6/14 17:21
 Desc: 百度指数
 感谢 https://cloudcrawler.club/categories/2019%E5%B9%B4%E6%9C%AB%E9%80%86%E5%90%91%E5%A4%8D%E4%B9%A0/
 """
@@ -50,16 +50,24 @@ def get_ptbk(uniqid: str, cookie: str) -> str:
         return ptbk
 
 
-def baidu_search_index(word: str = "python", start_date: str = "2020-01-01", end_date: str = "2020-05-01", cookie: str = None) -> str:
+def baidu_search_index(
+    word: str = "python",
+    start_date: str = "2020-01-01",
+    end_date: str = "2020-05-01",
+    cookie: str = None,
+    text: str = None,
+) -> str:
     headers = {
         "Accept": "application/json, text/plain, */*",
-        "Accept-Encoding": "gzip, deflate",
-        "Accept-Language": "zh-CN,zh;q=0.9",
+        "Accept-Encoding": "gzip, deflate, br",
+        "Accept-Language": "zh-CN,zh;q=0.9,en;q=0.8",
         "Cache-Control": "no-cache",
+        "Cipher-Text": text,
         "Cookie": cookie,
         "DNT": "1",
-        "Host": "zhishu.baidu.com",
+        "Host": "index.baidu.com",
         "Pragma": "no-cache",
+        "Referer": "https://index.baidu.com/v2/main/index.html",
         "Proxy-Connection": "keep-alive",
         "Referer": "zhishu.baidu.com",
         "User-Agent": "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_13_6) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/75.0.3770.90 Safari/537.36",
@@ -69,7 +77,7 @@ def baidu_search_index(word: str = "python", start_date: str = "2020-01-01", end
     session.headers.update(headers)
     params = {
         "area": "0",
-        "word": '[[{"name":' + f'"{word}"' + ',"wordType"' + ':1}]]',
+        "word": '[[{"name":' + f'"{word}"' + ',"wordType"' + ":1}]]",
         "startDate": start_date,
         "endDate": end_date,
     }
@@ -82,9 +90,14 @@ def baidu_search_index(word: str = "python", start_date: str = "2020-01-01", end
         ptbk = get_ptbk(uniqid, cookie)
         result = decrypt(ptbk, all_data).split(",")
         result = [int(item) if item != "" else 0 for item in result]
-        if len(result) == len(pd.date_range(start=start_date, end=end_date, freq="7D")):
+        if len(result) == len(
+            pd.date_range(start=start_date, end=end_date, freq="7D")
+        ):
             temp_df_7 = pd.DataFrame(
-                [pd.date_range(start=start_date, end=end_date, freq="7D"), result],
+                [
+                    pd.date_range(start=start_date, end=end_date, freq="7D"),
+                    result,
+                ],
                 index=["date", word],
             ).T
             temp_df_7.index = pd.to_datetime(temp_df_7["date"])
@@ -92,7 +105,10 @@ def baidu_search_index(word: str = "python", start_date: str = "2020-01-01", end
             return temp_df_7
         else:
             temp_df_1 = pd.DataFrame(
-                [pd.date_range(start=start_date, end=end_date, freq="1D"), result],
+                [
+                    pd.date_range(start=start_date, end=end_date, freq="1D"),
+                    result,
+                ],
                 index=["date", word],
             ).T
             temp_df_1.index = pd.to_datetime(temp_df_1["date"])
@@ -100,12 +116,15 @@ def baidu_search_index(word: str = "python", start_date: str = "2020-01-01", end
             return temp_df_1
 
 
-def baidu_info_index(word: str, start_date: str, end_date: str, cookie: str) -> str:
+def baidu_info_index(
+    word: str, start_date: str, end_date: str, cookie: str, text: str
+) -> str:
     headers = {
         "Accept": "application/json, text/plain, */*",
         "Accept-Encoding": "gzip, deflate",
         "Accept-Language": "zh-CN,zh;q=0.9",
         "Cache-Control": "no-cache",
+        "Cipher-Text": text,
         "Cookie": cookie,
         "DNT": "1",
         "Host": "zhishu.baidu.com",
@@ -119,12 +138,13 @@ def baidu_info_index(word: str, start_date: str, end_date: str, cookie: str) -> 
     session.headers.update(headers)
     params = {
         "area": "0",
-        "word": '[[{"name":' + f'"{word}"' + ',"wordType"' + ':1}]]',
+        "word": '[[{"name":' + f'"{word}"' + ',"wordType"' + ":1}]]",
         "startDate": start_date,
         "endDate": end_date,
     }
     with session.get(
-        url=f"http://index.baidu.com/api/FeedSearchApi/getFeedIndex", params=params
+        url=f"http://index.baidu.com/api/FeedSearchApi/getFeedIndex",
+        params=params,
     ) as response:
         data = response.json()["data"]
         all_data = data["index"][0]["data"]
@@ -132,9 +152,14 @@ def baidu_info_index(word: str, start_date: str, end_date: str, cookie: str) -> 
         ptbk = get_ptbk(uniqid, cookie)
         result = decrypt(ptbk, all_data).split(",")
         result = [int(item) if item != "" else 0 for item in result]
-        if len(result) == len(pd.date_range(start=start_date, end=end_date, freq="7D")):
+        if len(result) == len(
+            pd.date_range(start=start_date, end=end_date, freq="7D")
+        ):
             temp_df_7 = pd.DataFrame(
-                [pd.date_range(start=start_date, end=end_date, freq="7D"), result],
+                [
+                    pd.date_range(start=start_date, end=end_date, freq="7D"),
+                    result,
+                ],
                 index=["date", word],
             ).T
             temp_df_7.index = pd.to_datetime(temp_df_7["date"])
@@ -142,7 +167,10 @@ def baidu_info_index(word: str, start_date: str, end_date: str, cookie: str) -> 
             return temp_df_7
         else:
             temp_df_1 = pd.DataFrame(
-                [pd.date_range(start=start_date, end=end_date, freq="1D"), result],
+                [
+                    pd.date_range(start=start_date, end=end_date, freq="1D"),
+                    result,
+                ],
                 index=["date", word],
             ).T
             temp_df_1.index = pd.to_datetime(temp_df_1["date"])
@@ -150,12 +178,19 @@ def baidu_info_index(word: str, start_date: str, end_date: str, cookie: str) -> 
             return temp_df_1
 
 
-def baidu_media_index(word: str = "口罩", start_date: str = "2018-01-01", end_date: str = "2020-04-20", cookie: str = None) -> str:
+def baidu_media_index(
+    word: str = "口罩",
+    start_date: str = "2018-01-01",
+    end_date: str = "2020-04-20",
+    cookie: str = None,
+    text: str = None,
+) -> str:
     headers = {
         "Accept": "application/json, text/plain, */*",
         "Accept-Encoding": "gzip, deflate",
         "Accept-Language": "zh-CN,zh;q=0.9",
         "Cache-Control": "no-cache",
+        "Cipher-Text": text,
         "Cookie": cookie,
         "DNT": "1",
         "Host": "zhishu.baidu.com",
@@ -169,7 +204,7 @@ def baidu_media_index(word: str = "口罩", start_date: str = "2018-01-01", end_
     session.headers.update(headers)
     params = {
         "area": "0",
-        "word": '[[{"name":' + f'"{word}"' + ',"wordType"' + ':1}]]',
+        "word": '[[{"name":' + f'"{word}"' + ',"wordType"' + ":1}]]",
         "startDate": start_date,
         "endDate": end_date,
     }
@@ -183,9 +218,14 @@ def baidu_media_index(word: str = "口罩", start_date: str = "2018-01-01", end_
         result = decrypt(ptbk, all_data).split(",")
         result = ["0" if item == "" else item for item in result]
         result = [int(item) for item in result]
-        if len(result) == len(pd.date_range(start=start_date, end=end_date, freq="7D")):
+        if len(result) == len(
+            pd.date_range(start=start_date, end=end_date, freq="7D")
+        ):
             temp_df_7 = pd.DataFrame(
-                [pd.date_range(start=start_date, end=end_date, freq="7D"), result],
+                [
+                    pd.date_range(start=start_date, end=end_date, freq="7D"),
+                    result,
+                ],
                 index=["date", word],
             ).T
             temp_df_7.index = pd.to_datetime(temp_df_7["date"])
@@ -193,7 +233,10 @@ def baidu_media_index(word: str = "口罩", start_date: str = "2018-01-01", end_
             return temp_df_7
         else:
             temp_df_1 = pd.DataFrame(
-                [pd.date_range(start=start_date, end=end_date, freq="1D"), result],
+                [
+                    pd.date_range(start=start_date, end=end_date, freq="1D"),
+                    result,
+                ],
                 index=["date", word],
             ).T
             temp_df_1.index = pd.to_datetime(temp_df_1["date"])
@@ -203,17 +246,31 @@ def baidu_media_index(word: str = "口罩", start_date: str = "2018-01-01", end_
 
 if __name__ == "__main__":
     cookie = ""
-    data = baidu_search_index(
-        word="python", start_date="2020-01-01", end_date="2020-09-14", cookie=cookie
-    )
-    print(data)
+    text = ""
 
-    data = baidu_info_index(
-        word="口罩", start_date="2019-07-03", end_date="2020-09-21", cookie=cookie
+    baidu_search_index_df = baidu_search_index(
+        word="python",
+        start_date="2020-01-01",
+        end_date="2020-09-14",
+        cookie=cookie,
+        text=text,
     )
-    print(data)
+    print(baidu_search_index_df)
 
-    data = baidu_media_index(
-        word="金融科技", start_date="2020-01-01", end_date="2020-09-20", cookie=cookie
+    baidu_info_index_df = baidu_info_index(
+        word="口罩",
+        start_date="2019-07-03",
+        end_date="2022-09-21",
+        cookie=cookie,
+        text=text,
     )
-    print(data)
+    print(baidu_info_index_df)
+
+    baidu_media_index_df = baidu_media_index(
+        word="金融科技",
+        start_date="2020-01-01",
+        end_date="2022-09-20",
+        cookie=cookie,
+        text=text,
+    )
+    print(baidu_media_index_df)