+++ docs/changelog.md
@@ -29,6 +29,22 @@
 
 ## 更新说明
 
+1.6.17 fix: fix stock_new_a_spot_em interface
+
+    1. 修复 stock_new_a_spot_em 接口上市日期字段
+
+1.6.16 fix: fix baidu_search_index interface
+
+    1. 修复 baidu_search_index 接口，对该接口增加 text 字段
+
+1.6.15 fix: fix bond_zh_cov_info interface
+
+    1. 修复 bond_zh_cov_info 接口，修改接口参数及返回数据
+
+1.6.14 fix: fix stock_hsgt_board_rank_em interface
+
+    1. 修复 stock_hsgt_board_rank_em 接口，字段调整
+
 1.6.13 add: add stock_new_a_spot_em interface
 
     1. 新增 stock_new_a_spot_em 接口，获取新股行情数据
@@ -736,6 +752,14 @@
     3. 修改 air_quality_hist 接口的请求日期格式，从 '2022-01-01' 为 '20220101'
 
 ## 版本更新说明
+
+1.6.17 fix: fix stock_new_a_spot_em interface
+
+1.6.16 fix: fix baidu_search_index interface
+
+1.6.15 fix: fix bond_zh_cov_info interface
+
+1.6.14 fix: fix stock_hsgt_board_rank_em interface
 
 1.6.13 add: add stock_new_a_spot_em interface